Learned how to use Arrays.sort.
Kinda frustrating because I had the right base logic in iterating
through both arrays using nested for loops, but lacking the (sort)
method of already arranging both of the arrays - the code I had written
was useless because it didn’t account for duplicates.

The problem inherently needs duplicates.

There is definite redundancy within the program in its current form.