Completed brute force solution for TwoSum
I overthought it, thinking that hash map was a necessity and the
original solution given was only applicable if x and y were next to
each other.

Just need to take a step back and do things slower.  Solution is much
simpler, just iterate through array once, then based on the input
target int, find the difference, then go through the rest of the array
looking for the difference in value.