URLify, two methods
First method is what I first did, but I realize that even if you do
charAt(), you can only change the character by replacing it with
another char.  So in order to add the remaining ‘2’ and ‘0’, you would
essentially have to push all the elements down the array.

The second method is a much simpler, learned from Stack Overflow, built
in String method to scan a String, replace your target (in this case,
whitespace), with your new (“%20”).